requirements.txt
# Python dependencies
uwsgi
Flask
python-dotenv
Flask-SQLAlchemy
pandas